+++ lib/twilio.ts
@@ -4,6 +4,7 @@
  */
 import Connection from './twilio/connection';
 import Device from './twilio/device';
+import { Logger } from './twilio/log';
 import { PreflightTest } from './twilio/preflight/preflight' ;
 
-export { Connection, Device, PreflightTest };
+export { Connection, Device, PreflightTest, Logger };