+++ tests/integration/logger.ts
@@ -1,0 +1,8 @@
+import * as assert from 'assert';
+import { Logger } from '../../lib/twilio';
+
+describe('Logger', () => {
+  it('the exposed logger should be defined', () => {
+    assert(Logger);
+  });
+});