+++ CHANGELOG.md
@@ -1,5 +1,29 @@
-2.0 (In Progress)
-=================
+2.0.0 (In Progress)
+===================
+
+New Features
+------------
+
+### LogLevel Module
+
+The SDK now uses the [`loglevel`](https://github.com/pimterry/loglevel) module. This exposes several new features for the SDK, including the ability to intercept log messages with custom handlers and the ability to set logging levels after instantiating a `Device`. To get an instance of the `loglevel` `Logger` class used internally by the SDK, there are several options.
+
+If your project uses the `Twilio` window global:
+```js
+const logger = Twilio.Logger;
+...
+logger.setLogLevel('DEBUG');
+```
+
+Or if you use the `NPM` package in a JS or TS project:
+```ts
+import { Logger as TwilioClientLogger } from 'twilio-client.js`;
+...
+TwilioClientLogger.setLogLevel('DEBUG');
+```
+
+Please see the original [`loglevel`](https://github.com/pimterry/loglevel) project for more documentation on usage.
+
 
 Changes
 -------